backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
psycopg2-binary==2.9.1
mysql-connector-python==8.0.26 